gameplay and zobrist hashing working drafts. Network Architecture bones in place, working on policy head mapping.
